Queries and such fixed typo
Open the .sql files in Notepad or Notepad++.

Open MySQL workbench and go to Database->Connect to Datatbase. You'll need to import you password. Copy the "dbcreationquery" code into the query window in the middle. Go to Query->Execute (All or Selection).

Clear the window then do the same with the insertquery.txt file.

Then open up the python 2.7 file in IDLE and change the user and password to the main one of your server.